Feat: configuration files (#15)
* feat: create a config file for api

* feat: create a config file for websocket

* feat: create a config file for inlive channel

* feat: create a config file for webrtc

* feat: exports all internal config module

* fix: change to lowercase

* fix: remove config/config.js and update the path
