+++ packages/internal/config/channel.js
@@ -1,0 +1,15 @@
+/**
+ * @typedef Channel
+ * @property {string} base_url - The inLive stream channel base URL
+ */
+
+/**
+ * @type {Channel} - The channel configuration
+ */
+const channel = {
+  base_url: 'https://channel.inlive.app',
+}
+
+Object.freeze(channel)
+
+export { channel }
