+++ packages/internal/config/index.js
@@ -1,0 +1,4 @@
+export { api } from './api.js'
+export { channel } from './channel.js'
+export { webrtc } from './webrtc.js'
+export { websocket } from './websocket.js'
