+++ packages/internal/modules.js
@@ -1,2 +1,2 @@
 export { fetchHttp } from './fetch-http/fetch-http.js'
-export * as config from './config/config.js'
+export * as config from './config/index.js'
